Saved API URL to .env file. Moved fetching functionality to separate files
